Pull out Model to own file, add test class
